Fix PageHeader tests
